adjusting bac according to the user's detail
Changed the instructions according to our final product.
Added some modifications regarding the user's BAC according to the user's age, gender and bmi.
Side note, when the user registers as "Others", there should be a prompt saying that we cannot rely solely on the user's age and BMI to modify the accurate BAC value and make the user acknowledge that fact. However it might cause an issue with the consent of those that don't associate being male or female.
Just throwing some ideas around:)